third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-targetY-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'targetY' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('targetY') is "2"
PASS convolveMatrixElement.getAttribute('targetY') is "0"
PASS successfullyParsed is true

TEST COMPLETE
